feat(core): Defined AccountId to hold a contract of being valid [WIP]
