+++ Cargo.toml
@@ -69,6 +69,9 @@ neard = { path = "./neard"}
 
 testlib = { path = "./test-utils/testlib" }
 
+[patch.crates-io]
+borsh = { git = 'https://github.com/near/borsh', branch = 'feature/extended-box-support' }
+
 [profile.release]
 lto = true        # Enable full link-time optimization.
 codegen-units = 1 # Use only 1 codegen-unit to enable full optimizations.
