+++ core/primitives/Cargo.toml
@@ -12,6 +12,7 @@ byteorder = "1.3"
 chrono = { version = "0.4.4", features = ["serde"] }
 derive_more = "0.99.3"
 easy-ext = "0.2"
+thiserror = "1"
 sha2 = "0.8"
 lazy_static = "1.4"
 serde = { version = "1", features = ["derive"] }
