+++ core/primitives/benches/serialization.rs
@@ -1,3 +1,5 @@
+use std::convert::TryInto;
+
 #[macro_use]
 extern crate bencher;
 
@@ -24,10 +26,10 @@ fn create_transaction() -> SignedTransaction {
     SignedTransaction::new(
         Signature::empty(KeyType::ED25519),
         Transaction {
-            signer_id: "123213123123".to_string(),
+            signer_id: "123213123123".try_into().unwrap(),
             public_key: PublicKey::empty(KeyType::ED25519),
             nonce: 123,
-            receiver_id: "1231231232131".to_string(),
+            receiver_id: "1231231232131".try_into().unwrap(),
             block_hash: Default::default(),
             actions,
         },
@@ -45,7 +47,7 @@ fn create_block() -> Block {
         1_000,
         CryptoHash::default(),
     );
-    let signer = InMemoryValidatorSigner::from_random("".to_string(), KeyType::ED25519);
+    let signer = InMemoryValidatorSigner::from_random("test".try_into().unwrap(), KeyType::ED25519);
     Block::produce(
         PROTOCOL_VERSION,
         genesis.header(),
