+++ core/primitives/src/account_id.rs
@@ -1,0 +1,349 @@
+use std::convert::TryFrom;
+
+use regex::Regex;
+
+pub const MIN_ACCOUNT_ID_LEN: usize = 2;
+pub const MAX_ACCOUNT_ID_LEN: usize = 64;
+
+lazy_static::lazy_static! {
+    /// See NEP#0006
+    static ref VALID_ACCOUNT_ID: Regex =
+        Regex::new(r"^(([a-z\d]+[\-_])*[a-z\d]+\.)*([a-z\d]+[\-_])*[a-z\d]+$").unwrap();
+    /// Represents a part of an account ID with a suffix of as a separator `.`.
+    static ref VALID_ACCOUNT_PART_ID_WITH_TAIL_SEPARATOR: Regex =
+        Regex::new(r"^([a-z\d]+[\-_])*[a-z\d]+\.$").unwrap();
+    /// Represents a top level account ID.
+    static ref VALID_TOP_LEVEL_ACCOUNT_ID: Regex =
+        Regex::new(r"^([a-z\d]+[\-_])*[a-z\d]+$").unwrap();
+
+    pub static ref SYSTEM_ACCOUNT: AccountId = AccountId::try_from("system").unwrap();
+}
+
+/// Account identifier. Provides access to user's state.
+#[derive(
+    Debug,
+    derive_more::Display,
+    Clone,
+    PartialEq,
+    Eq,
+    Hash,
+    derive_more::AsRef,
+    derive_more::Into,
+    serde::Serialize,
+    serde::Deserialize,
+    borsh::BorshSerialize,
+    borsh::BorshDeserialize,
+)]
+#[as_ref(forward)]
+pub struct AccountId(Box<str>);
+
+impl From<AccountId> for String {
+    fn from(account_id: AccountId) -> Self {
+        account_id.0.into()
+    }
+}
+
+impl AccountId {
+    pub fn len(&self) -> usize {
+        self.0.len()
+    }
+
+    pub fn validate(account_id: &str) -> Result<(), ParseAccountError> {
+        if account_id.len() < MIN_ACCOUNT_ID_LEN {
+            Err(ParseAccountError::TooShort)
+        } else if account_id.len() > MAX_ACCOUNT_ID_LEN {
+            Err(ParseAccountError::TooLong)
+        } else if !VALID_ACCOUNT_ID.is_match(account_id) {
+            Err(ParseAccountError::Invalid)
+        } else {
+            Ok(())
+        }
+    }
+
+    pub fn is_valid_top_level_account_id(&self) -> bool {
+        VALID_TOP_LEVEL_ACCOUNT_ID.is_match(self.as_ref())
+    }
+
+    /// Returns true if the signer_id can create a direct sub-account with the given account Id.
+    /// It assumes the signer_id is a valid account_id
+    pub fn is_valid_sub_account_id(&self, parent_account_id: &AccountId) -> bool {
+        if parent_account_id.len() >= self.len() {
+            return false;
+        }
+        // Will not panic, since valid account id is utf-8 only and the length is checked above.
+        // e.g. when `near` creates `aa.near`, it splits into `aa.` and `near`
+        let (prefix, suffix) = self.0.split_at(self.len() - parent_account_id.len());
+        if suffix != parent_account_id.as_ref() {
+            return false;
+        }
+        VALID_ACCOUNT_PART_ID_WITH_TAIL_SEPARATOR.is_match(prefix)
+    }
+}
+
+#[derive(Debug, thiserror::Error)]
+pub enum ParseAccountError {
+    #[error("the value is too long for account ID")]
+    TooLong,
+    #[error("the value is too short for account ID")]
+    TooShort,
+    #[error("the value has invalid characters for account ID")]
+    Invalid,
+}
+
+// NOTE: We cannot implement it over generic AsRef<str> due to an overlapping
+// implementation with std:
+// https://www.reddit.com/r/rust/comments/c1yubv/why_this_code_does_not_compile/
+impl std::convert::TryFrom<&str> for AccountId {
+    type Error = ParseAccountError;
+
+    fn try_from(account_id: &str) -> Result<Self, Self::Error> {
+        Self::validate(&account_id)?;
+        Ok(Self(account_id.into()))
+    }
+}
+
+impl std::convert::TryFrom<String> for AccountId {
+    type Error = ParseAccountError;
+
+    fn try_from(account_id: String) -> Result<Self, Self::Error> {
+        Self::validate(&account_id)?;
+        Ok(Self(account_id.into()))
+    }
+}
+
+impl std::convert::TryFrom<&[u8]> for AccountId {
+    type Error = ParseAccountError;
+
+    fn try_from(account_id: &[u8]) -> Result<Self, Self::Error> {
+        Self::try_from(std::str::from_utf8(account_id).map_err(|_| Self::Error::Invalid)?)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    const OK_ACCOUNT_IDS: &[&str] = &[
+        "aa",
+        "a-a",
+        "a-aa",
+        "100",
+        "0o",
+        "com",
+        "near",
+        "bowen",
+        "b-o_w_e-n",
+        "b.owen",
+        "bro.wen",
+        "a.ha",
+        "a.b-a.ra",
+        "system",
+        "over.9000",
+        "google.com",
+        "illia.cheapaccounts.near",
+        "0o0ooo00oo00o",
+        "alex-skidanov",
+        "10-4.8-2",
+        "b-o_w_e-n",
+        "no_lols",
+        "0123456789012345678901234567890123456789012345678901234567890123",
+        // Valid, but can't be created
+        "near.a",
+    ];
+
+    #[test]
+    fn test_is_valid_account_id() {
+        for account_id in OK_ACCOUNT_IDS {
+            assert!(
+                AccountId::validate(account_id).is_ok(),
+                "Valid account id {:?} marked invalid",
+                account_id
+            );
+        }
+
+        let bad_account_ids = vec![
+            "a",
+            "A",
+            "Abc",
+            "-near",
+            "near-",
+            "-near-",
+            "near.",
+            ".near",
+            "near@",
+            "@near",
+            "неар",
+            "@@@@@",
+            "0__0",
+            "0_-_0",
+            "0_-_0",
+            "..",
+            "a..near",
+            "nEar",
+            "_bowen",
+            "hello world",
+            "abcdefghijklmnopqrstuvwxyz.abcdefghijklmnopqrstuvwxyz.abcdefghijklmnopqrstuvwxyz",
+            "01234567890123456789012345678901234567890123456789012345678901234",
+            // `@` separators are banned now
+            "some-complex-address@gmail.com",
+            "sub.buy_d1gitz@atata@b0-rg.c_0_m",
+        ];
+        for account_id in bad_account_ids {
+            assert!(
+                AccountId::validate(&account_id.to_string()).is_err(),
+                "Invalid account id {:?} marked valid",
+                account_id
+            );
+        }
+    }
+
+    #[test]
+    fn test_is_valid_top_level_account_id() {
+        let ok_top_level_account_ids = vec![
+            "aa",
+            "a-a",
+            "a-aa",
+            "100",
+            "0o",
+            "com",
+            "near",
+            "bowen",
+            "b-o_w_e-n",
+            "0o0ooo00oo00o",
+            "alex-skidanov",
+            "b-o_w_e-n",
+            "no_lols",
+            "0123456789012345678901234567890123456789012345678901234567890123",
+        ];
+        for account_id in ok_top_level_account_ids {
+            let account_id = AccountId::try_from(account_id).unwrap();
+            assert!(
+                AccountId::is_valid_top_level_account_id(&account_id),
+                "Valid top level account id {:?} marked invalid",
+                account_id
+            );
+        }
+
+        let bad_top_level_account_ids = vec![
+            "near.a",
+            "b.owen",
+            "bro.wen",
+            "a.ha",
+            "a.b-a.ra",
+            "some-complex-address@gmail.com",
+            "sub.buy_d1gitz@atata@b0-rg.c_0_m",
+            "over.9000",
+            "google.com",
+            "illia.cheapaccounts.near",
+            "10-4.8-2",
+            "a",
+            "A",
+            "Abc",
+            "-near",
+            "near-",
+            "-near-",
+            "near.",
+            ".near",
+            "near@",
+            "@near",
+            "неар",
+            "@@@@@",
+            "0__0",
+            "0_-_0",
+            "0_-_0",
+            "..",
+            "a..near",
+            "nEar",
+            "_bowen",
+            "hello world",
+            "abcdefghijklmnopqrstuvwxyz.abcdefghijklmnopqrstuvwxyz.abcdefghijklmnopqrstuvwxyz",
+            "01234567890123456789012345678901234567890123456789012345678901234",
+            // Valid regex and length, but reserved
+            "system",
+        ];
+        for account_id in bad_top_level_account_ids {
+            let account_id = AccountId::try_from(account_id).unwrap();
+            assert!(
+                !AccountId::is_valid_top_level_account_id(&account_id),
+                "Invalid top level account id {:?} marked valid",
+                account_id
+            );
+        }
+    }
+
+    #[test]
+    fn test_is_valid_sub_account_id() {
+        let ok_pairs = vec![
+            ("test", "a.test"),
+            ("test-me", "abc.test-me"),
+            ("gmail.com", "abc.gmail.com"),
+            ("gmail.com", "abc-lol.gmail.com"),
+            ("gmail.com", "abc_lol.gmail.com"),
+            ("gmail.com", "bro-abc_lol.gmail.com"),
+            ("g0", "0g.g0"),
+            ("1g", "1g.1g"),
+            ("5-3", "4_2.5-3"),
+        ];
+        for (signer_id, sub_account_id) in ok_pairs {
+            let signer_id = AccountId::try_from(signer_id).unwrap();
+            let sub_account_id = AccountId::try_from(sub_account_id).unwrap();
+            assert!(
+                sub_account_id.is_valid_sub_account_id(&signer_id),
+                "Failed to create sub-account {:?} by account {:?}",
+                sub_account_id,
+                signer_id
+            );
+        }
+
+        let bad_pairs = vec![
+            ("test", ".test"),
+            ("test", "test"),
+            ("test", "est"),
+            ("test", ""),
+            ("test", "st"),
+            ("test5", "ббб"),
+            ("test", "a-test"),
+            ("test", "etest"),
+            ("test", "a.etest"),
+            ("test", "retest"),
+            ("test-me", "abc-.test-me"),
+            ("test-me", "Abc.test-me"),
+            ("test-me", "-abc.test-me"),
+            ("test-me", "a--c.test-me"),
+            ("test-me", "a_-c.test-me"),
+            ("test-me", "a-_c.test-me"),
+            ("test-me", "_abc.test-me"),
+            ("test-me", "abc_.test-me"),
+            ("test-me", "..test-me"),
+            ("test-me", "a..test-me"),
+            ("gmail.com", "a.abc@gmail.com"),
+            ("gmail.com", ".abc@gmail.com"),
+            ("gmail.com", ".abc@gmail@com"),
+            ("gmail.com", "abc@gmail@com"),
+            ("test", "a@test"),
+            ("test_me", "abc@test_me"),
+            ("gmail.com", "abc@gmail.com"),
+            ("gmail@com", "abc.gmail@com"),
+            ("gmail.com", "abc-lol@gmail.com"),
+            ("gmail@com", "abc_lol.gmail@com"),
+            ("gmail@com", "bro-abc_lol.gmail@com"),
+            ("gmail.com", "123456789012345678901234567890123456789012345678901234567890@gmail.com"),
+            (
+                "123456789012345678901234567890123456789012345678901234567890",
+                "1234567890.123456789012345678901234567890123456789012345678901234567890",
+            ),
+            ("aa", "ъ@aa"),
+            ("aa", "ъ.aa"),
+        ];
+        for (signer_id, sub_account_id) in bad_pairs {
+            let signer_id = AccountId::try_from(signer_id).unwrap();
+            let sub_account_id = AccountId::try_from(sub_account_id).unwrap();
+            assert!(
+                !sub_account_id.is_valid_sub_account_id(&signer_id),
+                "Invalid sub-account {:?} created by account {:?}",
+                sub_account_id,
+                signer_id
+            );
+        }
+    }
+}
