+++ core/primitives/src/receipt.rs
@@ -6,12 +6,12 @@ use serde::{Deserialize, Serialize};
 
 use near_crypto::{KeyType, PublicKey};
 
+use crate::account_id::{AccountId, SYSTEM_ACCOUNT};
 use crate::hash::CryptoHash;
 use crate::logging;
 use crate::serialize::{option_base64_format, u128_dec_format_compatible};
 use crate::transaction::{Action, TransferAction};
-use crate::types::{AccountId, Balance};
-use crate::utils::system_account;
+use crate::types::Balance;
 
 /// Receipts are used for a cross-shard communication.
 /// Receipts could be 2 types (determined by a `ReceiptEnum`): `ReceiptEnum::Action` of `ReceiptEnum::Data`.
@@ -45,12 +45,12 @@ impl Receipt {
     /// allowance of the access key. For gas refunds use `new_gas_refund`.
     pub fn new_balance_refund(receiver_id: &AccountId, refund: Balance) -> Self {
         Receipt {
-            predecessor_id: system_account(),
+            predecessor_id: SYSTEM_ACCOUNT.clone(),
             receiver_id: receiver_id.clone(),
             receipt_id: CryptoHash::default(),
 
             receipt: ReceiptEnum::Action(ActionReceipt {
-                signer_id: system_account(),
+                signer_id: SYSTEM_ACCOUNT.clone(),
                 signer_public_key: PublicKey::empty(KeyType::ED25519),
                 gas_price: 0,
                 output_data_receivers: vec![],
@@ -72,7 +72,7 @@ impl Receipt {
         signer_public_key: PublicKey,
     ) -> Self {
         Receipt {
-            predecessor_id: system_account(),
+            predecessor_id: SYSTEM_ACCOUNT.clone(),
             receiver_id: receiver_id.clone(),
             receipt_id: CryptoHash::default(),
