+++ core/primitives/src/types.rs
@@ -5,14 +5,13 @@ use serde::{Deserialize, Serialize};
 use near_crypto::PublicKey;
 
 use crate::account::{AccessKey, Account};
+pub use crate::account_id::AccountId;
 use crate::challenge::ChallengesResult;
 use crate::errors::EpochError;
 use crate::hash::CryptoHash;
 use crate::serialize::u128_dec_format;
 use crate::trie_key::TrieKey;
 
-/// Account identifier. Provides access to user's state.
-pub type AccountId = String;
 /// Hash used by a struct implementing the Merkle tree.
 pub type MerkleHash = CryptoHash;
 /// Validator identifier in current group.
