+++ core/primitives/src/validator_signer.rs
@@ -1,3 +1,4 @@
+use std::convert::TryFrom;
 use std::path::Path;
 use std::sync::Arc;
 
@@ -63,8 +64,9 @@ pub trait ValidatorSigner: Sync + Send {
 
 /// Test-only signer that "signs" everything with 0s.
 /// Don't use in any production or code that requires signature verification.
-#[derive(Default)]
+#[derive(smart_default::SmartDefault)]
 pub struct EmptyValidatorSigner {
+    #[default(AccountId::try_from("test").unwrap())]
     account_id: AccountId,
 }
 
@@ -110,7 +112,7 @@ impl ValidatorSigner for EmptyValidatorSigner {
 
     fn sign_account_announce(
         &self,
-        _account_id: &String,
+        _account_id: &AccountId,
         _peer_id: &PeerId,
         _epoch_id: &EpochId,
     ) -> Signature {
@@ -138,17 +140,13 @@ pub struct InMemoryValidatorSigner {
 
 impl InMemoryValidatorSigner {
     pub fn from_random(account_id: AccountId, key_type: KeyType) -> Self {
-        Self {
-            account_id: account_id.clone(),
-            signer: Arc::new(InMemorySigner::from_random(account_id, key_type)),
-        }
+        let signer = Arc::new(InMemorySigner::from_random(account_id.clone().into(), key_type));
+        Self { account_id, signer }
     }
 
-    pub fn from_seed(account_id: &str, key_type: KeyType, seed: &str) -> Self {
-        Self {
-            account_id: account_id.to_string(),
-            signer: Arc::new(InMemorySigner::from_seed(account_id, key_type, seed)),
-        }
+    pub fn from_seed(account_id: AccountId, key_type: KeyType, seed: &str) -> Self {
+        let signer = Arc::new(InMemorySigner::from_seed(account_id.as_ref(), key_type, seed));
+        Self { account_id, signer }
     }
 
     pub fn public_key(&self) -> PublicKey {
@@ -157,7 +155,10 @@ impl InMemoryValidatorSigner {
 
     pub fn from_file(path: &Path) -> Self {
         let signer = InMemorySigner::from_file(path);
-        Self { account_id: signer.account_id.clone(), signer: Arc::new(signer) }
+        Self {
+            account_id: AccountId::try_from(signer.account_id.as_str()).unwrap(),
+            signer: Arc::new(signer),
+        }
     }
 }
